+++ src/main/java/org/a_sply/porter/dao/interfaces/MessageListDao.java
@@ -5,5 +5,5 @@ import java.util.List;
 import org.a_sply.porter.domain.MessageList;
 
 public interface MessageListDao {
-	List<MessageList> findByUserId(int id);
+	List<MessageList> findByUserId(long id);
 }
